extract common benchmark helpers
